CIF-1039 - Refactor CartProvider
