+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/__test__/cartOptions.test.js
@@ -15,7 +15,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 
 import CartOptions from '../cartOptions';
-import { CartProvider } from '../../../utils/state';
+import { CartProvider } from '../cartContext';
 
 describe('<CartOptions>', () => {
     it('renders the component properly', () => {
