+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/__test__/product.test.js
@@ -14,7 +14,7 @@
 import React from 'react';
 import Product from '../product';
 import { render } from '@testing-library/react';
-import { CartProvider } from '../../../utils/state';
+import { CartProvider } from '../cartContext';
 
 const mockCartItem = {
     product: {
