+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/__test__/productlist.test.js
@@ -14,7 +14,7 @@
 import React from 'react';
 import ProductList from '../productList';
 import { render } from 'test-utils';
-import { CartProvider } from '../../../utils/state';
+import { CartProvider } from '../cartContext';
 
 describe('<ProductList>', () => {
     const mockCartItems = [
