+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/cartContext.js
@@ -14,7 +14,7 @@
 
 import React, { createContext, useContext, useReducer } from 'react';
 import { object, func } from 'prop-types';
-import { useCookieValue } from '../utils/hooks';
+import { useCookieValue } from '../../utils/hooks';
 
 export const initialState = {
     isOpen: false,
