+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/minicart.js
@@ -26,7 +26,7 @@ import classes from './minicart.css';
 import CART_DETAILS_QUERY from '../../queries/query_cart_details.graphql';
 import CartTrigger from '../CartTrigger';
 
-import { useCartState } from '../../utils/state';
+import { useCartState } from './cartContext';
 
 const MiniCart = () => {
     const [{ cartId, cart, isOpen, isLoading, isEditing, addItem, errorMessage }, dispatch] = useCartState();
